refactor: global providers
